+++ ts/src/lib/index.ts
@@ -2,6 +2,7 @@ import { ReadableStream } from "@open-flash/stream";
 import { Action } from "avm1-tree/action";
 import { UintSize } from "semantic-types";
 import { ActionHeader, parseAction, parseActionHeader } from "./parsers/avm1";
+import { ParseError } from "./parsers/parse-error";
 
 export { cfgFromBytes } from "./cfg-from-bytes";
 
@@ -16,7 +17,7 @@ export class Avm1Parser {
     return this.stream.bytePos;
   }
 
-  readNext(): Action | undefined {
+  readNext(): Action | ParseError | undefined {
     if (this.stream.bytePos === this.stream.byteEnd) {
       return undefined;
     } else if (this.stream.peekUint8() === 0) {
@@ -26,7 +27,7 @@ export class Avm1Parser {
     return parseAction(this.stream);
   }
 
-  readAt(offset: UintSize): Action | undefined {
+  readAt(offset: UintSize): Action | ParseError | undefined {
     this.stream.bytePos = offset;
     return this.readNext();
   }